Switch to using dedicated sketch compilation action repository
The `arduino/actions/libraries/compile-examples` GitHub Actions action has been moved to a dedicated repository. The
action at the previous experimental arduino/actions repository is no longer being maintained.

There was a breaking change made to in the default sketches report folder name in the action since it was moved to the
dedicated repository. Rather than making the minimal change and continuing to rely on the new default folder name, the
report folder is now explicitely defined in the workflow.